requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
channels==4.0.0
channels-redis==4.1.0
Django==4.2
djangorestframework==3.14.0
msgpack==1.0.5
pytz==2023.3
redis==4.5.4
sqlparse==0.4.3
tzdata==2023.3
